requirements.txt
requests
pandas
beautifulsoup4
vaderSentiment
google-generativeai
